move levels to config
