+++ vllm/compilation/backends.py
@@ -13,10 +13,9 @@ import vllm.envs as envs
 from vllm.logger import init_logger
 from vllm.utils import combine_fx_passes, weak_ref_tensors
 
-from .config import CompilationConfig
+from .config import CompilationConfig, CompilationLevel
 from .counter import compilation_counter
 from .fusion import FusionPass
-from .levels import CompilationLevel
 from .reshapes import RedundantReshapesPass
 
 logger = init_logger(__name__)
