+++ vllm/v1/worker/gpu_model_runner.py
@@ -10,8 +10,7 @@ import torch.nn as nn
 
 from vllm import envs
 from vllm.compilation.compile_context import set_compile_context
-from vllm.compilation.config import CompilationConfig
-from vllm.compilation.levels import CompilationLevel
+from vllm.compilation.config import CompilationConfig, CompilationLevel
 from vllm.config import VllmConfig
 from vllm.forward_context import set_forward_context
 from vllm.inputs import INPUT_REGISTRY, InputRegistry
